Migrate ResetPasswordSuccess component to SetNewPassword
